Added auc.R
Calculates area under the curve.  At the moment, compares to x axis
(,0) so requires x and y vectors to be transformed before use.  x and y
are vectors in which a y observation is at an x value, but x values do
not need to be sorted. The function sorts x and y before calculating auc